Add self host
Self host helps to test the changes locally.
