+++ network_host.py
@@ -182,10 +182,11 @@ if __name__ == "__main__":
     parser.add_argument("--led-rgb-sequence", action="store", help="Switch if your matrix has led colors swapped. Default: RGB", default="RGB", type=str)
     parser.add_argument("--led-row-addr-type", action="store", help="0 = default; 1=AB-addressed panels", default=0, type=int, choices=[0,1])
     parser.add_argument("--led-multiplexing", action="store", help="Multiplexing type: 0=direct; 1=strip; 2=checker; 3=spiral (Default: 0)", default=0, type=int, choices=[0,1,2,3])
-
     args = parser.parse_args()
 
-    efficient_tracker = Efficient(args)
-    server = EfficientServer((args.host, args.port), EfficientHandler, efficient_tracker)
+    display = LedDisplay(args)
+    efficient = Efficient(display)
+    server = EfficientServer((args.host, args.port), EfficientHandler, efficient)
     log.info("Efficient server started at port {0}".format(args.port))
+
     server.serve_forever()
